Extract terminal output scheduler subsystems
